% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_praise.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(persons_name = "Nick", punctuation = "!")
}
\arguments{
\item{persons_name}{The name of the person you want to praise. Character vector of length one.}

\item{punctuation}{THe sign at the end of the sentence}
}
\value{
Text string with praise
}
\description{
This is a full description of the praise delivery function
}
