% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{text string; This is the name of the person you want to praise}

\item{punctuation}{text string; This is emphasis after you're name. Default is an exclamation mark.}
}
\value{
Text string with praise
}
\description{
This function is used to deliver praise when you're sad.
}
\examples{
praise(name = "Aidan", punctuation = "!")
}
